Use the canonical value of `undefined` in generated JavaScript code.
Using `undefined` in generated code is fragile as it is global mutable
variable, not a constant. As a consequence a buggy equality test such
as `if (foo = undefined) { ... }` or similar in generated code or any
third-party code can inadvertently break the entire runtime. Such bugs
can be difficult to localise.

To make the runtime more robust this patch introduces a slight change
to the JavaScript code generator such that the emitted code uses the
canonical value of `undefined` rather than the variable name
`undefined`. The canonical value can be obtain via the odd expression
`(void 0)`. Since `void` is a keyword and `0` is a numeric literal it
is not possible for user-code to override their semantics, hence the
expression is resilient against mutations such as monkey-patching.