Use getExecutablePath in getInstallDir
Also removes now obsolete getProgramPath.

getExecutablePath is a more robust implementation of getProgramPath
which is shipped with base. It also should work on Windows, although
testing would be appreciated.

The main advantage is that getExecutablePath has not the issue that it
wouldn't come up with a path to the executable if it isn't in PATH or
invoked in an unusual manner (e. g. via a symlink). getProgramPath would
then sometimes just return "koka" (getProgName exactly) and even worse,
dirname p in getInstallDir would be "". getExecutablePath guarantees in
such cases that we have an absolute path to work with.