Clarify doc for with-handlers
As written, there was not enough context for me to understand the section, reading the tour from the top.  In describing the desugaring, it seemed to be explaining `with handler` in terms of `handler`, for which there was no explanation.  I couldn't make head or tail of the section until got to the very last sentence.

I took the liberty of changing the `emit-console` names to `hello-console`, which IMO better describes the semantics and will help to guide the reader along; I hope you don't mind!